# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Dependency Ghost API - detect silent behavior changes in dependency upgrades
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface ScanInput {
  /** GitHub repository URL to scan */
  repoUrl: string;
  /**
     * GitHub sandbox repo (owner/repo) to run workflows in
     * @nullable
     */
  sandboxRepo?: string | null;
}

export type ScanStatus = typeof ScanStatus[keyof typeof ScanStatus];


export const ScanStatus = {
  pending: 'pending',
  thinking: 'thinking',
  executing: 'executing',
  correcting: 'correcting',
  completed: 'completed',
  failed: 'failed',
} as const;

export interface Scan {
  id: number;
  repoUrl: string;
  /** @nullable */
  sandboxRepo?: string | null;
  status: ScanStatus;
  /**
     * JSON contract map from Agent 1 (THINK)
     * @nullable
     */
  contractMap?: string | null;
  /**
     * GitHub Actions run ID from Agent 2 (EXECUTE)
     * @nullable
     */
  workflowRunId?: string | null;
  /**
     * Pull Request URL from Agent 3 (SELF-CORRECT)
     * @nullable
     */
  prUrl?: string | null;
  /** @nullable */
  errorMessage?: string | null;
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
}

export type ScanLogAgent = typeof ScanLogAgent[keyof typeof ScanLogAgent];


export const ScanLogAgent = {
  THINK: 'THINK',
  EXECUTE: 'EXECUTE',
  CORRECT: 'CORRECT',
  SYSTEM: 'SYSTEM',
} as const;

export type ScanLogLevel = typeof ScanLogLevel[keyof typeof ScanLogLevel];


export const ScanLogLevel = {
  info: 'info',
  warning: 'warning',
  error: 'error',
  success: 'success',
} as const;

export interface ScanLog {
  id: number;
  scanId: number;
  agent: ScanLogAgent;
  level: ScanLogLevel;
  message: string;
  createdAt: string;
}

export type MismatchSeverity = typeof MismatchSeverity[keyof typeof MismatchSeverity];


export const MismatchSeverity = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  critical: 'critical',
} as const;

/**
 * @nullable
 */
export type MismatchPatchStatus = typeof MismatchPatchStatus[keyof typeof MismatchPatchStatus] | null;


export const MismatchPatchStatus = {
  pending: 'pending',
  verified: 'verified',
  failed: 'failed',
  null: 'null',
} as const;

export interface Mismatch {
  id: number;
  scanId: number;
  dependency: string;
  functionName: string;
  expected: string;
  actual: string;
  severity: MismatchSeverity;
  /**
     * Generated compatibility wrapper code
     * @nullable
     */
  patch?: string | null;
  /** @nullable */
  patchStatus?: MismatchPatchStatus;
  createdAt: string;
}

export interface ScanStats {
  totalScans: number;
  completedScans: number;
  totalMismatches: number;
  fixedMismatches: number;
  recentScans: Scan[];
}

export interface SandboxSetupInput {
  /** Name for the sandbox repository */
  repoName: string;
}

export interface SandboxInfo {
  repoUrl: string;
  repoFullName: string;
  workflowPath: string;
  /** @nullable */
  message?: string | null;
}

export interface WorkflowYaml {
  yaml: string;
}

export interface WebhookConfig {
  /** Slack or Discord incoming webhook URL */
  url: string;
  /** Send notification when a scan completes successfully */
  notifyOnComplete: boolean;
  /** Send notification when a scan fails */
  notifyOnFailure: boolean;
}

export type WebhookTestResultProvider = typeof WebhookTestResultProvider[keyof typeof WebhookTestResultProvider];


export const WebhookTestResultProvider = {
  slack: 'slack',
  discord: 'discord',
  generic: 'generic',
} as const;

export interface WebhookTestResult {
  ok: boolean;
  provider: WebhookTestResultProvider;
  /** @nullable */
  statusCode?: number | null;
  /** @nullable */
  error?: string | null;
}

export type DeleteWebhookConfig200 = {
  ok: boolean;
};