# lib/api-zod/src/generated/types/webhookConfig.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Dependency Ghost API - detect silent behavior changes in dependency upgrades
 * OpenAPI spec version: 0.1.0
 */

export interface WebhookConfig {
  /** Slack or Discord incoming webhook URL */
  url: string;
  /** Send notification when a scan completes successfully */
  notifyOnComplete: boolean;
  /** Send notification when a scan fails */
  notifyOnFailure: boolean;
}