# lib/api-zod/src/generated/types/webhookTestResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Dependency Ghost API - detect silent behavior changes in dependency upgrades
 * OpenAPI spec version: 0.1.0
 */
import type { WebhookTestResultProvider } from './webhookTestResultProvider';

export interface WebhookTestResult {
  ok: boolean;
  provider: WebhookTestResultProvider;
  /** @nullable */
  statusCode?: number | null;
  /** @nullable */
  error?: string | null;
}